refactor: change from http put requests to patch requests to minimize the amount of data transfer

+++ server/src/main/java/com/unihub/app/dto/community/content/response/CalendarEventResponseDto.java
@@ -13,7 +13,7 @@ public record CalendarEventResponseDto(
         String title,
         EventType type,
         OffsetDateTime startTime,
-        Double durationHours,
+        Float durationHours,
         EventLocation location,
         String courseAbbreviation,
         boolean isSubscribed
